venchrent/requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fake-useragent==0.1.11
filelock==3.7.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
glob2==0.7
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
packaging==21.3
parse==1.19.0
parse-type==0.6.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyee==8.2.2
PyMySQL==1.0.2
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.1.2
pytest-bdd==5.0.0
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
svgwrite==1.4.2
tomli==2.0.1
tqdm==4.64.0
Tree==0.2.4
urllib3==1.26.9
virtualenv==20.14.1
w3lib==1.22.0
websockets==10.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
